renamed main to core_api.py
