--- main.py
@@ -1,261 +1,0 @@
-#!/usr/bin/env python3
-"""
-Main entry point for SAR Ship Detection Project.
-
-This module orchestrates the key functionalities including data acquisition,
-preprocessing, ship detection, tracking, and output generation through
-various user interfaces.
-"""
-
-import logging
-from typing import Dict, List, Optional, Tuple, Any
-from pathlib import Path
-import json
-from preprocessing.noise_filter import apply_correction
-from preprocessing.Land_masking import process_image, compare_images
-import cv2
-import numpy as np
-
-class ConfigurationManager:
-    """Handles system parameters and model configurations."""
-    
-    def __init__(self, config_path: Optional[str] = None):
-        pass
-    
-    def load_config(self) -> Dict[str, Any]:
-        pass
-    
-    def get_model_config(self) -> Dict[str, Any]:
-        pass
-    
-    def get_preprocessing_config(self) -> Dict[str, Any]:
-        pass
-
-
-class DataHandler:
-    """Manages data acquisition, temporary storage, and cleanup."""
-    
-    def __init__(self, config: Dict[str, Any]):
-        pass
-    
-    def download_from_copernicus(self, coordinates: Tuple[float, float], 
-                                time_range: Tuple[str, str]) -> List[str]:
-        pass
-    
-    def handle_user_upload(self, uploaded_file) -> str:
-        pass
-    
-    def load_preloaded_images(self) -> List[str]:
-        pass
-    
-    def cleanup_temporary_files(self, session_id: str) -> None:
-        pass
-
-
-'''class SARPreprocessor:
-    """Encapsulates the entire SAR image preprocessing pipeline."""
-    
-    def __init__(self, config: Dict[str, Any]):
-        pass
-    
-
-    
-    def apply_noise_reduction(self, image_data) -> Any:
-        pass
-    
-    def apply_land_masking(self, image_data) -> Any:
-        pass
-    
-    def standardize_format(self, image_data) -> Any:
-        pass'''
-
-
-class ShipDetector:
-    """Handles YOLOv11m inference for ship detection."""
-    
-    def __init__(self, model_path: str, config: Dict[str, Any]):
-        pass
-    
-    def load_model(self) -> None:
-        pass
-    
-    def detect_ships(self, preprocessed_image) -> Dict[str, Any]:
-        pass
-    
-    def extract_detections(self, model_output) -> List[Dict[str, Any]]:
-        pass
-
-
-class AISDetector:
-    """Handles AIS data integration for ship detection."""
-    
-    def __init__(self, ais_data_path: str):
-        pass
-    
-    def load_ais_data(self) -> List[Dict[str, Any]]:
-        pass
-    
-    def match_ais_with_detections(self, detections: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
-        pass
-
-
-
-class OutputGenerator:
-    """Generates various output formats and visualizations."""
-    
-    def __init__(self, config: Dict[str, Any]):
-        pass
-    
-    def generate_output_json(self, detections: List[Dict[str, Any]], 
-                           metadata: Dict[str, Any]) -> str:
-        pass
-    
-    def generate_annotated_image(self, original_image, 
-                               detections: List[Dict[str, Any]]) -> Any:
-        pass
-    
-    def extract_individual_ships(self, image, 
-                               detections: List[Dict[str, Any]]) -> List[Any]:
-        pass
-    
-    def generate_statistics(self, detections: List[Dict[str, Any]]) -> Dict[str, Any]:
-        pass
-
-
-class ProcessingQueue:
-    """Manages FIFO queue for processing requests."""
-    
-    def __init__(self, max_concurrent: int = 3):
-        pass
-    
-    def add_request(self, request_data: Dict[str, Any]) -> str:
-        pass
-    
-    def get_queue_position(self, request_id: str) -> int:
-        pass
-    
-    def process_next_request(self) -> Optional[Dict[str, Any]]:
-        pass
-    
-    def get_estimated_wait_time(self, request_id: str) -> int:
-        pass
-
-
-class StreamlitInterface:
-    """Manages the Streamlit web application and its different modes."""
-    
-    def __init__(self, config: Dict[str, Any]):
-        pass
-    
-    def run_streamlit_app(self) -> None:
-        pass
-    
-    def jetson_nano_interface(self) -> None:
-        pass
-    
-    def web_map_interface(self) -> None:
-        pass
-    
-    def image_upload_interface(self) -> None:
-        pass
-
-
-class SARShipDetectionSystem:
-    """Main application class orchestrating the entire workflow."""
-    
-    def __init__(self):
-        self.config_manager = None
-        self.data_handler = None
-        self.preprocessor = None
-        self.detector = None
-        self.tracker = None
-        self.output_generator = None
-        self.processing_queue = None
-        self.ui_interface = None
-        
-    def initialize_system(self) -> None:
-        pass
-    
-    def setup_logging(self) -> None:
-        pass
-    
-    def process_single_image(self, image_path: str, 
-                           session_id: str) -> Dict[str, Any]:
-        image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
-        if image is None:
-            raise FileNotFoundError(f"Could not load image at {image_path}")
-
-        image_new = apply_correction(image, image_path=image_path)
-        # Preprocess image (land masking)
-        masked_image, land_mask = process_image(image, visualize=False)
-
-        return {
-            "masked_image": masked_image,
-            "land_mask": land_mask,
-           }
-
-    
-    def process_image_sequence(self, image_paths: List[str], session_id: str) -> Dict[str, Any]:
-        results = []
-        for image_path in image_paths:
-            image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
-            if image is None:
-                results.append({
-                    "image_path": image_path,
-                    "error": f"Could not load image at {image_path}"
-                })
-                continue
-
-            # Optional: apply noise correction if needed
-            image_new = apply_correction(image, image_path=image_path)
-
-            # Preprocess image (land masking)
-            masked_image, land_mask = process_image(image, visualize=False)
-
-            results.append({
-                "image_path": image_path,
-                "masked_image": masked_image,
-                "land_mask": land_mask
-            })
-
-        return {
-            "session_id": session_id,
-            "results": results
-        }
-   
-    
-    def handle_error(self, error: Exception, context: str) -> None:
-        pass
-
-
-def setup_logging(log_level: str = "INFO") -> None:
-    """Configure logging for the application."""
-    pass
-
-
-def validate_environment() -> bool:
-    """Validate that all required dependencies and models are available."""
-    pass
-
-
-def main():
-    """Main entry point for the SAR Ship Detection application."""
-    try:
-        setup_logging()
-        
-        if not validate_environment():
-            logging.error("Environment validation failed")
-            return
-        
-        system = SARShipDetectionSystem()
-        system.initialize_system()
-        
-        system.ui_interface.run_streamlit_app()
-        
-    except Exception as e:
-        logging.error(f"Application failed to start: {e}")
-        raise
-
-
-if __name__ == "__main__":
-    main()
